Jenkinsfile: uncheck monitoring
- monitoring setting MUST be at 'most' what master instance is

(if master does not have monitoring, nor can slave)

... and then i rechecked Monitoring for both
Master AND Slave.

I tested it by manually starting the Agent with the
Jenkins UI. It failed when it was failing on CI build,
and now it succeeds, and I can see the new instances
in the EC2 dashboard.

I expect this to work.

Signed-off-by: meows <b5c6@protonmail.com>