core/vm: cache the instruction set
This chases an idea that generating a new instruction
set for each EVM interpreter may not be
very efficient.

It implements a basic cache that assumes that the
chain config, once set, won't change (which is currently
the case, since geth supposes that the chain config
is static once established at startup.

Date: 2021-03-10 11:11:48-06:00
Signed-off-by: meows <b5c6@protonmail.com>