core: fix total difficulty assignment to block from genesis
In core/genesis.go a default value for difficulty
is provided in GenesisToBlock, the vars.GenesisDifficulty value.

This value was not used in CommitGenesis, where instead
genesis.Difficulty (instead of block.Difficulty()) was used.

As you can see in the diff, the vars.GenesisDifficulty.Int64() value
is added to the preexisting expected TD outcomes for Short and Long reorg
tests.

Date: 2021-10-21 13:00:04-07:00
Signed-off-by: meows <b5c6@protonmail.com>